refactor: rename spectator_toggle.js to smart_gamemode.ts
