+++ src/bp/scripts/commands/smart_gamemode.ts
