+++ src/bp/scripts/index.ts
@@ -7,5 +7,5 @@ import "./commands/help.js";
 import "./commands/kill_item.js";
 import "./commands/move_forward.js";
 import "./commands/nv_toggle.js";
-import "./commands/spectator_toggle.js";
+import "./commands/smart_gamemode.js";
 import "./commands/user_custom_commands.js";
